% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioread.R
\name{format_channel}
\alias{format_channel}
\title{Convert a python channel to something better for R}
\usage{
format_channel(pych)
}
\arguments{
\item{pych}{A channel in Python format}
}
\value{
A channel in R format
}
\description{
Convert a python channel to something better for R
}
\keyword{internal}
