% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioread.R
\name{format_marker}
\alias{format_marker}
\title{Convert a python marker to a better R format}
\usage{
format_marker(pym)
}
\arguments{
\item{pym}{A marker in Python format}
}
\value{
A marker in R format
}
\description{
Convert a python marker to a better R format
}
\keyword{internal}
