% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioread.R
\name{read_acq}
\alias{read_acq}
\title{Read a BIOPAC AcqKnowledge file}
\usage{
read_acq(file, channel_indexes = NULL)
}
\arguments{
\item{file}{Path to an .acq file}

\item{channel_indexes}{Specific channel indexes to load data for (NULL loads all channels)}
}
\value{
A named list containing .acq data in R format with components:
\describe{
\item{channels}{A list of channel data and metadata}
\item{markers}{A list of markers}
\item{journal}{A character vector of journal text}
}
}
\description{
Read a BIOPAC AcqKnowledge file
}
\details{
The \code{channels} component is a list of channels, each containing:
\describe{
\item{time}{A vector of time points}
\item{data}{A vector of channel data}
\item{name}{The channel name}
\item{units}{The channel units}
\item{samples_per_second}{The number of samples per second}
}

The \code{markers} component is a list of markers, each containing:
\describe{
\item{time_index}{The time index of the marker, in seconds}
\item{sample_index}{The sample index of the marker}
\item{text}{The marker text}
\item{channel_number}{The channel number of the marker}
\item{channel_name}{The channel name of the marker}
\item{type}{The type of marker}
\item{color_rgba}{The color of the marker, as a 4-element vector of RGBA byte values}
\item{date_created}{The date and time the marker was created, in ISO 8601 format}
}

Some marker components are NULL for some file versions.
}
\examples{
acq_file = system.file("extdata", "physio-5.0.1-c.acq", package = "bioread")
acq_data <- read_acq(acq_file)

channel <- acq_data$channels[[1]]

# Plot the first channel
plot(
    channel$time, channel$data, type = "l", xlab="Time (s)",
    ylab = channel$units, main = channel$name)

}
